+++ kubedb.com/v1alpha2/redis.go
@@ -93,7 +93,7 @@ func (r Redis) modeFn(obj map[string]interface{}) (string, error) {
 
 func (r Redis) roleResourceFn(fn func(rr core.ResourceRequirements) core.ResourceList) func(obj map[string]interface{}) (map[api.PodRole]core.ResourceList, error) {
 	return func(obj map[string]interface{}) (map[api.PodRole]core.ResourceList, error) {
-		exporter, err := api.ContainerResources(obj, fn, "spec", "monitor", "prometheus")
+		exporter, err := api.ContainerResources(obj, fn, "spec", "monitor", "prometheus", "exporter")
 		if err != nil {
 			return nil, err
 		}
