+++ testdata/kubedb.com/v1alpha2/mysql/standalone.yaml
@@ -4,7 +4,16 @@ metadata:
   name: mysql
   namespace: demo
 spec:
-  version: "8.0.23-v1"
+  version: 8.0.23
+  podTemplate:
+    spec:
+      resources:
+        requests:
+          cpu: 250m
+          memory: 500Mi
+        limits:
+          cpu: 500m
+          memory: 1Gi
   storageType: Durable
   storage:
     storageClassName: "standard"
@@ -19,9 +28,9 @@ spec:
       exporter:
         resources:
           requests:
-            cpu: 100m
-            memory: 128Mi
+            cpu: 250m
+            memory: 64Mi
           limits:
-            cpu: 100m
+            cpu: 500m
             memory: 128Mi
   terminationPolicy: WipeOut
