--- testdata/kubedb.com/v1alpha2/redis.yaml
@@ -1,15 +1,0 @@
-apiVersion: kubedb.com/v1alpha2
-kind: Redis
-metadata:
-  name: redis-quickstart
-  namespace: demo
-spec:
-  version: 6.0.6
-  storageType: Durable
-  storage:
-    storageClassName: "standard"
-    accessModes:
-    - ReadWriteOnce
-    resources:
-      requests:
-        storage: 1Gi
