+++ testdata/kubedb.com/v1alpha2/redis/standalone.yaml
@@ -1,13 +1,24 @@
 apiVersion: kubedb.com/v1alpha2
-kind: MongoDB
+kind: Redis
 metadata:
-  name: mongodb
+  name: redis-quickstart
   namespace: demo
 spec:
-  version: "4.2.3"
+  version: 6.0.6
+  podTemplate:
+    spec:
+      resources:
+        requests:
+          cpu: 250m
+          memory: 500Mi
+        limits:
+          cpu: 500m
+          memory: 1Gi
   storageType: Durable
   storage:
     storageClassName: "standard"
+    accessModes:
+    - ReadWriteOnce
     resources:
       requests:
         storage: 1Gi
@@ -17,9 +28,8 @@ spec:
       exporter:
         resources:
           requests:
-            cpu: 100m
-            memory: 128Mi
+            cpu: 250m
+            memory: 64Mi
           limits:
-            cpu: 100m
+            cpu: 500m
             memory: 128Mi
-  terminationPolicy: WipeOut
